Fix conditionals for visibility for input box and progress (#169053)
I think this is left over from a previous PR of mine https://github.com/microsoft/vscode/pull/122054/files

Previously if there were no items in the quick pick, the input box would be forced. Additionally, if there was no input box, no progress would be shown.

Both of this don't make a ton of sense from the user's perspective so I've updated that to:

* hide the input if it's been asked to be hidden
* show the progress bar regardless of whether the input is shown

The only exception to the second rule is if there is no input and no description which means we're in a `Ctrl`+`Tab` quick behavior.

Fixes #167554