Add default keybindings for shell integration
These new keybindings will help with the discoverability of shell
integration features, especially run recent command when accessibility
mode is enabled. The alt+f1 menu in #170977 will help with
discoverability further and also help with the discovery of ctrl+alt+r.

When in accessibility mode:
- ctrl+r: run recent command
- ctrl+alt+r: send ctrl+r

When not in accessibility mode:
- ctrl+r: send ctrl+r
- ctrl+alt+r: run recent command

Regardless of accessibility mode:
- ctrl+g: go to recent directory
- ctrl+alt+g: send ctrl+g

Fixes #170976