ui: reduce transition of status bar background color (#172270)
When testing a new feature to disable debug treatment of the status bar,
I found that debugging felt _much_ more snappy when there wasn't a big
slow color fade on the status bar. Rob agreed it was a bit "dramatic."

A 150ms fade instead of a 350ms fade feels much nicer, I think we should
try this out for next month's Insiders.