Unfreeze Git Commit Message grammar II (#173195)
* Unfreeze Git Commit Message grammar II

Before this change, the upstream for the VSCode Git grammar was dead.
Also, the test setup for that project has been EOL since 2014, so even
just running the tests was difficult.

The replacement grammar, unlike the current grammar:
* Has a vscode-tmgrammar-test test suite that is runnable and passing
  and that will run in CI for any PRs (in the upstream project)
* Has diff highlighting for Swedish as well as English (#133888)
* Highlights touched files both in Swedish and in English

Fixes #133888
Fixes #168847

Ref: <https://github.com/walles/git-commit-message-plus>

And for the record, I was the one setting up the new Git Commit Message
project. And it was fun!

* Remedy review feedback

Retain the line-too-long subject line highlighting. Improved to
highlight only the too-long part, but same idea still.

Special case English language file operations keywords and retain the
previous classification of those. But fallback to op-and-filename
classification when that fails (like it will for Swedish git for
example).

* Update colorize test result

* Update script and cgmanifest

---------

Co-authored-by: Alex Ross <alros@microsoft.com>