testing: add a test results panel (#175414)
Previously, as in the issue this fixes #175377, messages without
location were treated a little strangely, and messages for tests without
a location were impossible to view. Them transient nature of the peek
has been a longstanding overall experience papercut.

This PR adds a Test Results panel which users can keep around as a
persistent results view. It shares the code used to generate the peek
view. With this PR, I've also tweaked it slightly so that the test
'history' is by default (unless toggled on) shown only in the panel,
aligning the test errors peek more towards the diagnostics view, where
the peek is a lighter, local view, and the panel shows everything.

The only real 'addition' to the previously-peek content in this PR is
that test output is shown when the top-level run node is selected.
_Ideally_ this would be in a terminal where full ANSI codes and colors
are supported, but it seems terminal land does not support this very
well, and I want to eventually get ANSI color support in text mode
anyway (#151964).