Handle `ParameterBindingException` for PowerShell 5.1
Which ships with PSReadLine 2.0.0 whose `Get-PSReadLineKeyHandler`
function does not support `-Chord`.
Easiest fix is catching the exception in that rare case and using
`-Bound` and a filter to do the same thing.

Note that most people do upgrade PSReadLine,
but there are cases where an inbox PowerShell 5.1 is being used
without an update (and it may not always be updateable).

Interestingly `Set-PSReadLineKeyHandler` does have `-Chord`
in this older version, so it can be left alone.