testing: tie test output correctly to runs, rework ui accordingly (#180746)
* testing: don't enqueue tests into tasks that are created

This is an old artifact that caused problems in the issue this fixes #180041

When we get a new 'task' from a test run request, we automatically
marked all tests as being enqueued with it. However, tests are now lazily
added to the test run states, so this would only affect tests that
_already_ had their state set to something (e.g. in another task). And
therefore it was also ineffective for its original purpose of marking
all included tests as being Queued, since extensions have had to do that
themselves anyway.

So just remove this code! Also, adjust priorities such that "skipped"
priorities are shown above unset ones.

* testing: tie test output correctly to runs, rework ui accordingly

Previously, test output was handled for an entire TestRunRequest into
a single stream. This didn't match what the public API implied, nor what
the real intentions were.

This makes output be stored correctly on a per-test run (called "tasks"
internally). It changes the order of the Test Results view so that tests
nested under their task, and also improves some of the tree handling
there to update nodes more specifically.

Clicking on the "terminal" button in the test view continues to show
output from the last test run request, but if there were multiple tasks,
the user is asked to choose which they want to view output for.

Finally I removed some old unused code that dealt with storing test
results on disk, which we no longer do, and moved to a simpler interface
instead.

Fixes #180041