Don't throw on main thread when messaging disposed webviews (#182132)
Fixes #173291

It's possible for the ext host to try performing an action on a webview that has been disposed but where the dispose message hasn't yet reached the ext host.  We currently throw in these cases and there's nothing an extension could do to prevent this

With this change, I've removed the exception logic so these types of messages to the main thread will just noop. On the ext host side, we still throw an error if you try posting to a webview that the ext host knows for sure is disposed of. This is usually a real bug in extension code. Extensions can prevent it by checking if the webview they are holding has been disposed of before making a request to it