Fix paste edit providers incorrectly overwriting data transfer values (#182169)
At present, copy providers on the ext host side return the full data transfer object passed to them (which includes any additions they make). We then use this to construct a new data transfer that is passed on to paste providers

There are a few bugs with this approach:

- If there are multiple copy providers, they can incorrectly end up overwriting each other. For example if the initial data transfer contains `text/plain` and there are two providers, the first of which tries writing a new `text/plain`, the second provider will overwrite this with the initial `text/plain` value

- If the original data transfer contains multiple entries for a mime, these always end up being overwritten with a single value

- We shouldn't waste type transferring value back to the main thread if the main thread already has them

This PR tries to fix this by skipping ext host to main thread transfer of non-modified data transfer items. As part of this work, I reworked a few internal structures and introduced a new `IReadonlyVSDataTransfer` interface to make it more clear when a internal data transfer is or is not expected to be modified