remote: add resource provider api for embedders (#181969)
Initially I was looking at VS Code providing some service worker code
to plop in in an opaque way for embedders. But that was convoluted both
for an API and consumers. For example, if we provided an agnostic handler
for this, it would clash with the Workbox types (a popular service worker
toolkit we use in vscode.dev) since they don't expose the original events.

Instead, take a more minimal approach, allowing embedders to hook in
themselves and provide a route that they want to use for the 'proxy'.

This was quite easy to implement as a variation in vscode.dev. Example
of extension icons loading:

![](https://memes.peet.io/img/23-05-8da09054-2717-4fca-b9b9-c605718f22f0.png)