testing: use overflow ellipsis in test tree (#186806)
* tree: add fast path for children updates in compressedObjectTreeModel

I noticed in testing that more things were getting rerendered than they
should have been. This adds a fast path so that, if the elements in
a compressed node don't change when its children are updated, we only
update the children instead of replacing the entire node from the parent
collection.

* testing: use overflow ellipsis in test tree

Fixes #186320

Had to abandon flexbox for this, so do manual tweaks to make codicons show nicely:

![](https://memes.peet.io/img/23-06-321bf892-8f69-4548-9c76-465f33205c8e.png)