testing: refactor to reduce explorer complexity (#186954)
Previously the two 'projections' in the test explorer, a list and tree
view, inherited from each other. This was a pretty messy way to do it.
There were other things with how we dealt with the tree that weren't
optimal, and some debt from concepts that no longer exist.

In this PR the now better-named ListProject and TreeProjection are
mostly indepenent from each other. This reduces the complexity in the
TestProjection and makes the ListProjection pretty simple.