fix: gutter testing spinner icon broken due to line height (#187081)
Getting the "wobble" entirely out was annoying. I initially assumed the countainer had to be either an odd or dimension for flexbox to center the icon, but this seemed to not be sufficient. For example, a line height of 36 and 40 were wobble-free, but 38 was not.

So instead take a sledgehammer at spinning icons and center them with a transform position.

Fixes #183269