testing: improve area below filter (#189441)
With results, the area is both more compact and useful. Passed/Total
tests are shown on the left, with an icon summarizing the view. On the
right, the wall-clock duration of the tests and a button to rerun tests.
The alt action on the button reruns the tests with the debugger.
Previously the rerun action was not very discoverable--had a few issues
about it over the course of time.

![](https://memes.peet.io/img/23-08-8bb85d55-6304-4acb-bef3-9c1126da8c49.png)

There's also no awkward gap when there's no test results:

![](https://memes.peet.io/img/23-08-f5bcda56-60d7-4c8d-9822-e311b68079d7.png)

Fixes #188841

cc @eleanorjboyd