Increase the target of Monaco from es6 to es2018
A target of es6 is equivalent to es2015. A target of es2017 adds support
for async / await. A target of es2018 adds support for `for await...of`.

Increasing the target means these syntaxes are no longer down-leveled.
This increases readability and debuggability of the compiled code and
potentially performance.

All modern browsers support this. https://caniuse.com/?search=await