tasks: use a type union for ITaskEvent (#192314)
When fixing a bug in tests I noticed that all properties of the
ITaskEvent were optional, so I had to go into task code to figure out
what types I could use.

This PR refactors the types so ITaskEvent is a union type that provides
as much information as possible about the presence of properties.

I tested this on some projects but am not sure if there's a repo/project
to test out the full gamut of task functionality. But there should be
no (barring once, see comment) functional change.