explorer: integrate test coverage bars (#198891)
* explorer: integrate test coverage bars

This adds coverage bars in the explorer view. These are only shown when
test coverage is open, and I took some care to avoid any performance
hit when it's not. The idle case involves no DOM manipulation and just
an observable listener on the test result service.

In the explorer, coverage is shown in its 'compact' view which will only
render a single bar representing the user's chosen overall test coverage
calculation (defaults to a weighted sum of all types of coverage
available).

Works with rename, decorations, and ellipses overflow. Screenshot:

For #123713

* rework explorer coverage bars to a contribution model