testing: add initial editor decorations
This is the first pass at decorations in-editor. This PR doesn't
actually register the contribution, as it's not ready for selfhosting
yet. This PR creates decorations that look like this. The idea is that
coverage decorations in the glyph margin will always be visibile when
there's coverage, and users can get coverage in their code via hover or
shortcut, with the intention of making coverage unobtrusive and easy to
run all the time.

![](https://memes.peet.io/img/24-01-8e61f4db-f115-4732-affe-59dea879a335.png)

The notable thing is that there is now a third glyph margin row. I
reworked some of the editor code to handle this.

![](https://memes.peet.io/img/24-01-f400369f-650c-4303-be65-e65903f8ad17.png)

Some open questions:

- The glyph margin coverage wants doesn't need to be full-width, should
  we add a new 'leftmost' glyph lane instead that's thinner?
- Adding breakpoints in files with coverage is a little annoying since
  the breakpoint hint widget can expand the glyph margin on lines with
	coverage, and jump back over otherwise. Probably we should never
	decrease the number of lanes shown whenever the cursor is over the
	glyph margin.

		![](https://memes.peet.io/img/24-01-79b53dd9-6fca-41dd-87b5-a113f9c25efb.gif)